==> 4_code/main/1_DATA_plant_preprocessing.ipynb <==
# 해당 파일 설명
이 코드를 모두 실행하고 나면, /convert 경로에 다음의 세 파일이 생성됩니다.

1. plant1_by_loc_with_datetime
2. plant2_by_loc_with_datetime
3. plant_test_with_datetime

## 추가 설명
위의 세 파일은, 날짜, plant, loc을 기반으로 제공 받은 데이터에 관한 feature를 정리하여 저장합니다.

# 1. Google drive Mount

from google.colab import auth
auth.authenticate_user()

from google.colab import drive
drive.mount('/content/gdrive')


# 2. Define Directory

# 사용하기 전 변경해주어야 하는 파라미터
BASE_WORKING_DIR = "/content/gdrive/My Drive/weather_bigcontest/4_code/"
#BASE_WORKING_DIR = "/content/gdrive/My Drive/Colab Notebooks/weather_bigcontest/4_code/"
USERNAME = "park"

import os
# 사용하는 디렉토리 정의
DATA_BASE_PATH = os.path.join(BASE_WORKING_DIR, "data")
DATA_RAW_PATH = os.path.join(DATA_BASE_PATH, "raw")
DATA_EXPORT_PATH = os.path.join(DATA_BASE_PATH, "export")
DATA_CONV_USER_PATH = os.path.join(DATA_BASE_PATH, USERNAME)
DATA_CONV_PATH = os.path.join(DATA_BASE_PATH, "converted")
MODEL_PATH = os.path.join(BASE_WORKING_DIR, "model")
RESULT_PATH = os.path.join(BASE_WORKING_DIR, "result")

WORKING_DIR = os.path.join(BASE_WORKING_DIR, USERNAME)

print('Current Work Directory :',  os.getcwd(), '\n')
print('Username :',  USERNAME, '\n')

print("DATA_BASE_PATH : ", DATA_BASE_PATH)
print("DATA_RAW_PATH : ", DATA_RAW_PATH)
print("DATA_EXPORT_PATH : ", DATA_EXPORT_PATH)
print("DATA_CONV_USER_PATH : ", DATA_CONV_USER_PATH)
print("DATA_CONV_PATH : ", DATA_CONV_PATH)

print("MODEL_PATH : ", MODEL_PATH)

print("RESULT_PATH : ", RESULT_PATH)

print("WORKING_DIR : ", WORKING_DIR)


# 데이터 무엇이 있는지 확인하자.
# 아래 코드를 수행하게 되면, "../input/"에서 리눅스 명령어 "ls"를 수행한 결과를 보여주게 된다. 이 때 나온 binary code를 "utf8"로 디코딩해서 보여준다.
from subprocess import check_output
print("===== Raw Data List =====\n", check_output(["ls", DATA_RAW_PATH]).decode("utf8"))
print("===== Export Data List =====\n" ,check_output(["ls", DATA_EXPORT_PATH]).decode("utf8"))
print("===== Converted User Data List =====\n" ,check_output(["ls", DATA_CONV_USER_PATH]).decode("utf8"))
print("===== Converted Data List =====\n" ,check_output(["ls", DATA_CONV_PATH]).decode("utf8"))
print("===== Model List =====\n" ,check_output(["ls", MODEL_PATH]).decode("utf8"))
print("===== Result file List =====\n" ,check_output(["ls", RESULT_PATH]).decode("utf8"))

# 4. Import Library

# 기본 데이터 라이브러리 로드
import sys # 시스템 파라미터에 접근할 수 있게 도와준다.
print("Python version: {}". format(sys.version))

import pandas as pd # 데이터 정제에 도움을 주는 라이브러리
print("pandas version: {}". format(pd.__version__))

import matplotlib # 매트랩에서 사용하는 시각화 도구를 사용할 수 있게 도와주는 시각화 도구
print("matplotlib version: {}". format(matplotlib.__version__))

import numpy as np # 행렬 계산을 위해 필요한 라이브러리
print("NumPy version: {}". format(np.__version__))

import scipy as sp # 수학 관련 함수가 내장된 라이브러리
print("SciPy version: {}". format(sp.__version__)) 

import IPython
from IPython import display # 주피터 노트북에서 예쁘게 시각화 해주는 도구
print("IPython version: {}". format(IPython.__version__)) 

import sklearn # 각종 통계 도구와 머신 러닝 알고리즘이 내장되어 있는 라이브러리
print("scikit-learn version: {}". format(sklearn.__version__))

# 파이썬 내장 라이브러리
import random
import datetime
import time
import os
from datetime import datetime, timedelta
from dateutil.relativedelta import relativedelta
import copy


# Jupyter Notebook 이나 ipython 을 사용하다보면 향후 버전이 올라갈 때 변경될 사항 등을 알려주는 경고 메시지(warning message)를 뜨지 않게 해준다.
import warnings
warnings.filterwarnings('ignore')
print('-'*25)

# 모델링 라이브러리, 수학 계산 툴 로드
# 일반적인 모델링 라이브러리
from sklearn import svm, tree, linear_model, neighbors, naive_bayes, ensemble, discriminant_analysis, gaussian_process
from xgboost import XGBClassifier

# 모델링 시 헬퍼 함수들
from sklearn.preprocessing import OneHotEncoder, LabelEncoder
from sklearn import feature_selection
from sklearn import model_selection
from sklearn import metrics

# interpolation 함수
import scipy.interpolate as spi

# model load
import pickle
from sklearn.externals import joblib

# 시각화 도구
import matplotlib as mpl
import matplotlib.pyplot as plt
import matplotlib.pylab as pylab
import seaborn as sns
from pandas.plotting import scatter_matrix

# 시각화 도구 default 세팅
# 주피터 노트북에서 plot 결과를 볼 수 있게 해준다.
%matplotlib inline
mpl.style.use('ggplot') # matplotlib에서 plot되는 결과를 선택할 수 있다.
sns.set_style('white') # seaborn에서 사용할 style을 설정할 수 있다.
pylab.rcParams['figure.figsize'] = 12,8 # plot의 크기와 선 등의 기본 값을 설정할 수 있다.

# 시각화 Style 설정하기
# print(plt.style.available)
# ['bmh', 'classic', 'dark_background', 'fast', 'fivethirtyeight', 'ggplot', 'grayscale', 'seaborn-bright', 'seaborn-colorblind', 'seaborn-dark-palette', 'seaborn-dark', 'seaborn-darkgrid', 'seaborn-deep', 'seaborn-muted', 'seaborn-notebook', 'seaborn-paper', 'seaborn-pastel', 'seaborn-poster', 'seaborn-talk', 'seaborn-ticks', 'seaborn-white', 'seaborn-whitegrid', 'seaborn', 'Solarize_Light2', 'tableau-colorblind10', '_classic_test']

plt.style.use(['fivethirtyeight'])
sns.set_style('darkgrid')

# 5. Read Raw Data

# 데이터 나누기
# 우리는 데이터를 3단계로 나눠야 한다.
# 1. train data
# 2. test data(만든 모델을 test)
# 3. validation(최종 예측 용)
data_raw_plant1 = pd.read_csv(os.path.join(DATA_RAW_PATH, "plant1_train.csv"))
data_raw_plant2 = pd.read_csv(os.path.join(DATA_RAW_PATH, "plant2_train.csv"))
data_val  = pd.read_csv(os.path.join(DATA_RAW_PATH, "plant_test.csv"))

# unnamed: 0 제거
data_raw_plant1.drop(['Unnamed: 0'], axis='columns', inplace=True)
data_raw_plant2.drop(['Unnamed: 0'], axis='columns', inplace=True)
data_val.drop(['Unnamed: 0'], axis='columns', inplace=True)

# column 명 변경
data_raw_plant1.columns = ['mea_ddhr', 'tem_in_loc1','hum_in_loc1', 'tem_coil_loc1','tem_in_loc2', 'hum_in_loc2','tem_coil_loc2', 'tem_in_loc3','hum_in_loc3', 'tem_coil_loc3','tem_out_loc1', 'hum_out_loc1','cond_loc1', 'cond_loc2','cond_loc3']
data_raw_plant2.columns = ['mea_ddhr', 'tem_in_loc1','hum_in_loc1', 'tem_coil_loc1','tem_in_loc2', 'hum_in_loc2','tem_coil_loc2', 'tem_in_loc3','hum_in_loc3', 'tem_coil_loc3','tem_out_loc1', 'hum_out_loc1','cond_loc1', 'cond_loc2','cond_loc3']
data_val.columns = ['mea_ddhr','plant','loc','tem_in','hum_in','tem_coil','tem_out','hum_out','x24h_tma','cond_loc24','x48h_tma','cond_loc48']

# 기본 데이터는 냅두고, 이걸 복사해서 쓰는 것이 효율적이다.
# 여기서 python이 기본적으로 reference로 shellow copy를 한다는 점을 인지하고 deep copy를 해야한다.
data1_plant1 = data_raw_plant1.copy(deep = True)
data1_plant2 = data_raw_plant2.copy(deep = True)

# copy by reference를 한다는 점을 이용하여, 빈번하게 수정할 데이터를 한번에 list로 갖고 있자.
data_cleaner = [data1_plant1, data1_plant2, data_val]

# 6. Simple Check



## 1. Check INFO

#preview data
print("========= 1공장 TRAIN INFO =========\n")
data_raw_plant1.info()
print("\n========= 2공장 TRAIN INFO =========\n")
data_raw_plant2.info()
print("\n========= VALIDATION INFO =========\n")
data_val.info()

print("========= 1공장 TRAIN SAMPLE =========\n")
data_raw_plant1.sample(10).transpose() # random으로 10개를 뽑아준다.

print("========= 2공장 TRAIN SAMPLE =========\n")
data_raw_plant2.sample(10).transpose() # random으로 10개를 뽑아준다.

print("========= VALIDATION SAMPLE =========\n")
data_val.sample(10).transpose() # random으로 10개를 뽑아준다.

## 2. Check NULL

def countNull(df):
  """
  describe : 데이터 갯수, NULL 갯수, 백분율을 보기위한 코드
  <input>
  pd.dataframe
  <output>
  pd.dataframe
  """
  count_null = pd.DataFrame()
  count_null['numOfTotal'] = [len(df) for x in df.columns]
  count_null['numOfNull'] = list(df.isnull().sum())
  count_null['numOfNotnull'] = count_null['numOfTotal'] - count_null['numOfNull']
  count_null['percent'] = count_null['numOfNull']/count_null['numOfTotal']*100
  count_null['dataType'] = [type(df[x][0]) for x in df.columns]
  count_null.index = df.columns
  return count_null

# plant1의 NULL 갯수
count_null_plant1 = countNull(data_raw_plant1)
count_null_plant1

# plant2의 NULL 갯수
count_null_plant2 = countNull(data_raw_plant2)
count_null_plant2

plant2 데이터를 보면 모든 변수의 널의 갯수가 동일함을 알수 있다.

따라서 같은 row에 모든 널값이 있는지 확인한다.

# 확인하는 코드
sum(data_raw_plant2.isnull().sum(axis=1) > 0)

결과적으로 null값은 모두 동일한 row에 분포 되어있음을 알수 있다.

즉 374개 정답값을 제외한 변수가 존재 하지 않음으로 제거하는것이 옳다.

## 3. Check Label Distribution

def checkLabel(df, plant):
  label = pd.DataFrame()
  label['true'] = [sum(df['cond_loc1'] == 1),
                   sum(df['cond_loc2'] == 1),
                   sum(df['cond_loc3'] == 1)]
  label['false'] = [sum(df['cond_loc1'] == 0),
                   sum(df['cond_loc2'] == 0),
                   sum(df['cond_loc3'] == 0)]
  label['percent'] = label['true'] / (label['true'] + label['false']) * 100
  label.index = ['loc1', 'loc2', 'loc3']
  return label

# plant1 loc에 대한 분포
label1 = checkLabel(data_raw_plant1, 1)
label1

# plant2 loc에 대한 분포
label2 = checkLabel(data_raw_plant2, 2)
label2

# 7. Data Arrange

## 1. 사용하는 함수

def arrangeData(df, plant):
  """
  describe : plant, loc에 따라 column을 볼 수 있게 데이터 프레임을 만든다.
  """
  ret_columns_name = ['mea_ddhr','plant','loc','tem_in','hum_in','tem_coil','tem_out_loc','hum_out_loc','cond_loc']
  ret_df = pd.DataFrame(columns=ret_columns_name)
  # plant = int(df.columns[1].split('_')[0][-1])
  np_df = np.array(df)
  ret_df = np.repeat(' '*30, len(df)*3*9).reshape(len(df)*3, 9)
  for index in range(len(df)):
    for loc in range(3):
      p = 3*loc + 1
      ret_df[3*index + loc,0] = np_df[index, 0]
      ret_df[3*index + loc,1] = plant
      ret_df[3*index + loc,2] = loc + 1
      ret_df[3*index + loc,3] = np_df[index, p] # tem_in
      ret_df[3*index + loc,4] = np_df[index, p+1] # hum_in
      ret_df[3*index + loc,5] = np_df[index, p+2] # tem_coil
      ret_df[3*index + loc,6] = np_df[index, 10] # tem_out
      ret_df[3*index + loc,7] = np_df[index, 11] # hum_out
      ret_df[3*index + loc,8] = np_df[index, 12+loc]
  ret_df = pd.DataFrame(ret_df, columns=ret_columns_name)
  for col in ret_columns_name:
    if col == 'mea_ddhr':
      ret_df[col] = ret_df[col].apply(lambda x : str(x) if x != 'nan' else np.nan)
    else:
      ret_df[col] = ret_df[col].apply(lambda x : np.float64(x) if x != 'nan' else np.nan)
  return ret_df

import datetime
def stringToDatetime(df, colName='mea_ddhr'):
  """
  describe : string 을 datetime으로 변경한다.
  input = pd.Dataframe
  output = pd.Dataframe
  """
  if type(df[colName][0]) != str: # 이미 자료형이 datetime인 경우 return 한다.
    return df
  if len(df[colName][0].split(' ')[1].split(':')) == 2: # 날짜 정보에 초에 대한 정보가 없을 경우 초를 붙여주고 datetime으로 변경한다.
    df[colName] = df[colName].apply(lambda x : x + ':00' if x != 'nan' else np.nan)
  df[colName] = df[colName].apply(lambda x : datetime.datetime.strptime(x, '%Y-%m-%d %H:%M:%S') if x != 'nan' else np.nan)
  return df

def createDateTime(df, colName='mea_ddhr'):
  """
  describe : datetime의 정보에서 date 정보와 0시 기준 몇 분인지를 나타내는 time을 분리하여 추가한다.
  input = pd.Dataframe
  output = pd.Dataframe
  """
  df = stringToDatetime(df, colName)
  df['date'] = df[colName].apply(lambda x : x.date())
  df['time'] = df[colName].apply(lambda x : np.float64(x.hour * 60 + x.minute))
  return df

## 2. 데이터 형식 변경

# 데이터 형식 변경
data_plant1 = arrangeData(data_raw_plant1, 1).copy(deep = True)
data_plant2 = arrangeData(data_raw_plant2, 2).copy(deep = True)

## 3. Date, time 추가

data_plant1 = createDateTime(data_plant1)
data_plant2 = createDateTime(data_plant2)

data_plant1

data_plant2

## 4. TEST DATA 형식 변경

tma24 = pd.DataFrame(data_val[['x24h_tma','plant','loc']])
tma24.columns = ['mea_ddhr','plant','loc']

tma48 = pd.DataFrame(data_val[['x48h_tma','plant','loc']])
tma48.columns = ['mea_ddhr','plant','loc']

data_test = pd.concat([tma24, tma48], ignore_index=True)
data_test = data_test.drop_duplicates()
data_test

data_test = createDateTime(data_test)
data_test

# 7. 결측치 다루기

is_NaN = data_plant1.isnull()
row_has_NaN = is_NaN.any(axis=1)
data_plant1[row_has_NaN]

## 1. Plant1에 대한 결측치 제거

먼저, 결측치가 함께 없는 row에 대해 제거를 수행한다.

### 1. Loc 1.0

# nan의 개수가 이상함을 감지.
a = data_plant1[data_plant1["cond_loc"].isnull()]
b = data_plant1[data_plant1["tem_in"].isnull()]
c = data_plant1[data_plant1["hum_in"].isnull()]
d = data_plant1[data_plant1["tem_coil"].isnull()]
a = a[a["loc"] == 1.0]
b = b[b["loc"] == 1.0]
c = c[c["loc"] == 1.0]
d = d[d["loc"] == 1.0]
print(a.shape[0])
print(b.shape[0])
print(c.shape[0])
print(d.shape[0])


# loc == 1.0일 때, tem_in, hum_in, tem_coil, cond_loc이 nan인 row는 같은 row이다. 검증하는 코드
a = data_plant1[["loc", "tem_in", "hum_in", "tem_coil", "cond_loc"]]
a = a[a["loc"] == 1.0]
a = a[["tem_in", "hum_in", "tem_coil", "cond_loc"]]
is_NaN = a.isnull()
row_has_NaN = is_NaN.all(axis=1)
print("870 * 4 = ", a[row_has_NaN].shape[0]*a[row_has_NaN].shape[1])



### 2. Loc 2.0

# nan의 개수가 이상함을 감지.
a = data_plant1[data_plant1["cond_loc"].isnull()]
b = data_plant1[data_plant1["tem_in"].isnull()]
c = data_plant1[data_plant1["hum_in"].isnull()]
d = data_plant1[data_plant1["tem_coil"].isnull()]
a = a[a["loc"] == 2.0]
b = b[b["loc"] == 2.0]
c = c[c["loc"] == 2.0]
d = d[d["loc"] == 2.0]
print(a.shape[0])
print(b.shape[0])
print(c.shape[0])
print(d.shape[0])


# loc == 1.0일 때, tem_in, hum_in, tem_coil, cond_loc이 nan인 row는 같은 row이다. 검증하는 코드
a = data_plant1[["loc", "tem_in", "hum_in", "tem_coil", "cond_loc"]]
a = a[a["loc"] == 2.0]
a = a[["tem_in", "hum_in", "tem_coil", "cond_loc"]]
is_NaN = a.isnull()
row_has_NaN = is_NaN.all(axis=1)
print("120 * 4 = ", a[row_has_NaN].shape[0]*a[row_has_NaN].shape[1])



### 3. Loc 3.0

# nan의 개수가 이상함을 감지.
a = data_plant1[data_plant1["cond_loc"].isnull()]
b = data_plant1[data_plant1["tem_in"].isnull()]
c = data_plant1[data_plant1["hum_in"].isnull()]
d = data_plant1[data_plant1["tem_coil"].isnull()]
a = a[a["loc"] == 3.0]
b = b[b["loc"] == 3.0]
c = c[c["loc"] == 3.0]
d = d[d["loc"] == 3.0]
print(a.shape[0])
print(b.shape[0])
print(c.shape[0])
print(d.shape[0])


tem_coil 가 nan인 개수가 보다 적다.

# loc == 1.0일 때, tem_in, hum_in, tem_coil, cond_loc이 nan인 row는 같은 row이다. 검증하는 코드
a = data_plant1[["loc", "tem_in", "hum_in", "tem_coil", "cond_loc"]]
a = a[a["loc"] == 3.0]
a = a[["tem_in", "hum_in", "tem_coil", "cond_loc"]]
is_NaN = a.isnull()
row_has_NaN = is_NaN.all(axis=1)
print("120 * 4 = ", a[row_has_NaN].shape[0]*a[row_has_NaN].shape[1])



### 4. cond_loc(y) 값이 없는 경우 제거

data_plant1.info()

# 그렇다면, cond_loc이 없는 경우에 대해 모두 삭제를 수행한다.
data_plant1 = data_plant1.dropna(subset = ["cond_loc"])
data_plant1.info()

## 2. Plant2에 대한 결측치 제거

# 그렇다면, cond_loc이 없는 경우에 대해 모두 삭제를 수행한다.
data_plant2 = data_plant2.dropna(subset = ["cond_loc"])
data_plant2.info()

y값이 없는 경우를 다 지우고 나니, 결측치가 삭제되었다.

# 8. 데이터 저장하기


# 정제된 데이터를 저장한다.
data_plant1.to_csv(os.path.join(DATA_CONV_PATH, 'plant1_by_loc_with_datetime.csv'))
data_plant2.to_csv(os.path.join(DATA_CONV_PATH, 'plant2_by_loc_with_datetime.csv'))
data_test.to_csv(os.path.join(DATA_CONV_PATH, 'plant_test_with_datetime.csv'))
