==> 4_code/main/16_MODEL_CLTR_INOUT_clustering.ipynb <==
# 해당 파일 설명

이 코드를 실행하기 위한 파일, /convert 경로에 다음의 파일이 필요합니다.

1. data_all_time_out_in.csv

이 코드를 모두 실행하고 나면, /converted 경로에 다음의 파일이 생성됩니다.

1. date_trend_point.csv

## 추가 설명

tem_in, tem_out, tem_coil, hum_in, hum_out 에 대한

point, trend 클러스터링을 적용하고 그결과를 date_trend_point.csv 저장합니다.

# 1. Google drive Mount

from google.colab import auth
auth.authenticate_user()

from google.colab import drive
drive.mount('/content/gdrive')

# 2. Define Directory

# 사용하기 전 변경해주어야 하는 파라미터
BASE_WORKING_DIR = "/content/gdrive/My Drive/weather_bigcontest/4_code/"
#BASE_WORKING_DIR = "/content/gdrive/My Drive/Colab Notebooks/weather_bigcontest/4_code/"
USERNAME = "park"

import os
# 사용하는 디렉토리 정의
DATA_BASE_PATH = os.path.join(BASE_WORKING_DIR, "data")
DATA_RAW_PATH = os.path.join(DATA_BASE_PATH, "raw")
DATA_EXPORT_PATH = os.path.join(DATA_BASE_PATH, "export")
DATA_CONV_USER_PATH = os.path.join(DATA_BASE_PATH, USERNAME)
DATA_CONV_PATH = os.path.join(DATA_BASE_PATH, "converted")
MODEL_PATH = os.path.join(BASE_WORKING_DIR, "model")
RESULT_PATH = os.path.join(BASE_WORKING_DIR, "result")

WORKING_DIR = os.path.join(BASE_WORKING_DIR, USERNAME)

print('Current Work Directory :',  os.getcwd(), '\n')
print('Username :',  USERNAME, '\n')

print("DATA_BASE_PATH : ", DATA_BASE_PATH)
print("DATA_RAW_PATH : ", DATA_RAW_PATH)
print("DATA_EXPORT_PATH : ", DATA_EXPORT_PATH)
print("DATA_CONV_USER_PATH : ", DATA_CONV_USER_PATH)
print("DATA_CONV_PATH : ", DATA_CONV_PATH)

print("MODEL_PATH : ", MODEL_PATH)

print("RESULT_PATH : ", RESULT_PATH)

print("WORKING_DIR : ", WORKING_DIR)


# 데이터 무엇이 있는지 확인하자.
# 아래 코드를 수행하게 되면, "../input/"에서 리눅스 명령어 "ls"를 수행한 결과를 보여주게 된다. 이 때 나온 binary code를 "utf8"로 디코딩해서 보여준다.
from subprocess import check_output
print("===== Raw Data List =====\n", check_output(["ls", DATA_RAW_PATH]).decode("utf8"))
print("===== Export Data List =====\n" ,check_output(["ls", DATA_EXPORT_PATH]).decode("utf8"))
print("===== Converted User Data List =====\n" ,check_output(["ls", DATA_CONV_USER_PATH]).decode("utf8"))
print("===== Converted Data List =====\n" ,check_output(["ls", DATA_CONV_PATH]).decode("utf8"))
print("===== Model List =====\n" ,check_output(["ls", MODEL_PATH]).decode("utf8"))
print("===== Result file List =====\n" ,check_output(["ls", RESULT_PATH]).decode("utf8"))

# 4. Import Library

# 기본 데이터 라이브러리 로드
import sys # 시스템 파라미터에 접근할 수 있게 도와준다.
print("Python version: {}". format(sys.version))

import pandas as pd # 데이터 정제에 도움을 주는 라이브러리
print("pandas version: {}". format(pd.__version__))

import matplotlib # 매트랩에서 사용하는 시각화 도구를 사용할 수 있게 도와주는 시각화 도구
print("matplotlib version: {}". format(matplotlib.__version__))

import numpy as np # 행렬 계산을 위해 필요한 라이브러리
print("NumPy version: {}". format(np.__version__))

import scipy as sp # 수학 관련 함수가 내장된 라이브러리
print("SciPy version: {}". format(sp.__version__)) 

import IPython
from IPython import display # 주피터 노트북에서 예쁘게 시각화 해주는 도구
print("IPython version: {}". format(IPython.__version__)) 

import sklearn # 각종 통계 도구와 머신 러닝 알고리즘이 내장되어 있는 라이브러리
print("scikit-learn version: {}". format(sklearn.__version__))

# 파이썬 내장 라이브러리
import random
import time
import os
from datetime import datetime, timedelta
from dateutil.relativedelta import relativedelta
import copy


# Jupyter Notebook 이나 ipython 을 사용하다보면 향후 버전이 올라갈 때 변경될 사항 등을 알려주는 경고 메시지(warning message)를 뜨지 않게 해준다.
import warnings
warnings.filterwarnings('ignore')
print('-'*25)

# 모델링 라이브러리, 수학 계산 툴 로드
# 일반적인 모델링 라이브러리
from sklearn import svm, tree, linear_model, neighbors, naive_bayes, ensemble, discriminant_analysis, gaussian_process
from xgboost import XGBClassifier
from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler  # 표준화 패키지 라이브러리 

# 모델링 시 헬퍼 함수들
from sklearn.preprocessing import OneHotEncoder, LabelEncoder
from sklearn import feature_selection
from sklearn import model_selection
from sklearn import metrics
from sklearn.datasets import make_blobs
from sklearn.cluster import KMeans

# interpolation 함수
import scipy.interpolate as spi

# model load
import pickle
from sklearn.externals import joblib

# 시각화 도구
import matplotlib as mpl
import matplotlib.pyplot as plt
import matplotlib.pylab as pylab
import seaborn as sns
from pandas.plotting import scatter_matrix

# 시각화 도구 default 세팅
# 주피터 노트북에서 plot 결과를 볼 수 있게 해준다.
%matplotlib inline
mpl.style.use('ggplot') # matplotlib에서 plot되는 결과를 선택할 수 있다.
sns.set_style('white') # seaborn에서 사용할 style을 설정할 수 있다.
pylab.rcParams['figure.figsize'] = 12,8 # plot의 크기와 선 등의 기본 값을 설정할 수 있다.

# 시각화 Style 설정하기
# print(plt.style.available)
# ['bmh', 'classic', 'dark_background', 'fast', 'fivethirtyeight', 'ggplot', 'grayscale', 'seaborn-bright', 'seaborn-colorblind', 'seaborn-dark-palette', 'seaborn-dark', 'seaborn-darkgrid', 'seaborn-deep', 'seaborn-muted', 'seaborn-notebook', 'seaborn-paper', 'seaborn-pastel', 'seaborn-poster', 'seaborn-talk', 'seaborn-ticks', 'seaborn-white', 'seaborn-whitegrid', 'seaborn', 'Solarize_Light2', 'tableau-colorblind10', '_classic_test']

plt.style.use(['fivethirtyeight'])
sns.set_style('darkgrid')

# 5. Get the Data

# date, plant, loc 별로 DataFrame 구축
cluster_df = pd.DataFrame(columns=['date', 'plant', 'loc'])

date_start = datetime.strptime("2016-04-01 00:00:00", '%Y-%m-%d %H:%M:%S').date()
date_end = datetime.strptime("2020-03-31 00:00:00", '%Y-%m-%d %H:%M:%S').date()
date = date_start

while date <= date_end:
  if date.month == 1 and date.day == 1:
      print(date)
  for plant in range(1, 3):
    for loc in range(1, 4):
      add_df = pd.DataFrame(columns=['date', 'plant', 'loc'])
      add_df.loc[0] = [date, plant, loc]
      cluster_df = cluster_df.append(add_df)
  date += timedelta(days=1)
cluster_df = cluster_df.reset_index()
cluster_df.drop(['index'], axis='columns', inplace=True)
cluster_df

# 혹시 몰라서 백업용
cluster_df_copy = copy.deepcopy(cluster_df)
cluster_df_copy

## date별 기상 + trend + point + in + out 불러오기

# date별 기상 + trend + point + in + out
data = pd.read_csv(os.path.join(DATA_CONV_PATH, "data_all_time_out_in.csv"))

data.drop(['Unnamed: 0'], axis='columns', inplace=True)

data = data[['date', 'plant',	'loc', 'time', 'tem_out_loc', 'hum_out_loc', 'tem_in', 'hum_in', 'tem_coil']]

data['date'] = data['date'].apply(lambda x : datetime.strptime(x + " 00:00:00", '%Y-%m-%d %H:%M:%S').date()) 

data

# IN, OUT 클러스터링

## 1. Trend clustering

# trend, point 클러스터링을 하기 위한 변수 이름
tem_name = ['tem_out_loc', 'hum_out_loc', 'tem_in', 'hum_in', 'tem_coil']

# trend 함수
import pickle
from sklearn.externals import joblib
def trend_fn(df, plant, loc, target, k=5):
  temp_trend = pd.DataFrame(columns=['date','0-4', '4-8','8-12','12-16','16-20','20-24'])

  date_start = datetime.strptime("2016-04-01 00:00:00", '%Y-%m-%d %H:%M:%S').date()
  date_end = datetime.strptime("2020-03-31 00:00:00", '%Y-%m-%d %H:%M:%S').date()
  date = date_start

  while date <= date_end:
    if date.month == 1 and date.day == 1:
      print(plant, loc, target, date)
    temp_date = pd.DataFrame(columns=['date','0-4', '4-8','8-12','12-16','16-20','20-24'])
    temp_date.loc[0] = 0
    temp_date['date'] = date
    for time in range(0, 24, 4):
      start_time = time * 60
      end_time = (time + 4) * 60
      if end_time == 24 * 60:
        end_time = 0
      start = df[(df['date'] == date) & (df['time'] == start_time)][target]
      end = df[(df['date'] == date) & (df['time'] == end_time)][target]
      temp_date['{}-{}'.format(time, time + 4)] = start[start.index[0]] - end[end.index[0]]
    temp_trend = temp_trend.append(temp_date)  # append를 사용하면 concat이랑 같은 효과임
    date += timedelta(days=1)
  #
  # 각 일에 대한 interpolation 함수를 대변하는 feature를 기반으로 클러스터링을 진행한다.
  kmeans = KMeans(n_clusters=k)
  kmeans.fit(temp_trend[['0-4', '4-8','8-12','12-16','16-20','20-24']])
  temp_trend[target + '-trend'] = kmeans.predict(temp_trend[['0-4', '4-8','8-12','12-16','16-20','20-24']])
  temp_trend['plant'] = plant
  temp_trend['loc'] = loc
  #print(temp_trend[['date', 'plant', 'loc', target + '-trend']])
  return temp_trend[['date', 'plant', 'loc', target + '-trend']]

# trend 만들기
for tem in tem_name:
  cluster_df[tem + '-trend'] = -1
for tem in tem_name:
  data_tem = data[['date', 'plant', 'loc', 'time', tem]]
  for plant in range(1, 3):
    for loc in range(1, 4):
      data_slice = data_tem[(data_tem['plant'] == plant) & (data_tem['loc'] == loc)]
      data_slice = trend_fn(data_slice, plant, loc, tem)
      temp_cluster_df = cluster_df[(cluster_df['plant'] == plant) & (cluster_df['loc'] == loc)]
      temp_cluster_df[tem + '-trend'] =  list(data_slice[tem + '-trend'])
      cluster_df.loc[temp_cluster_df.index] = temp_cluster_df
cluster_df

cluster_df_copy = copy.deepcopy(cluster_df)

## 2. Point Clustering

# point 함수
from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler

def point_fn(df, plant, loc, target, k=5):
  timeList = list(range(24))
  timeList.insert(0, "min")
  timeList.insert(0, "max")
  timeList.insert(0, "date")
  timeList = list(map(str, timeList))

  temp_point = pd.DataFrame(columns=timeList)

  date_start = datetime.strptime("2016-04-01 00:00:00", '%Y-%m-%d %H:%M:%S').date()
  date_end = datetime.strptime("2020-03-31 00:00:00", '%Y-%m-%d %H:%M:%S').date()
  date = date_start

  while date <= date_end:
    tem_df = df[df['date'] == date]
    if date.month == 1 and date.day == 1:
      print(plant, loc, target)
    
    temp_date = pd.DataFrame(columns=timeList)
    temp_date.loc[0] = 0
    temp_date['date'] = date
    # Max, Min column 넣어주기
    temp_date["min"] = np.min(list(tem_df[target]))
    temp_date["max"] = np.max(list(tem_df[target]))

    # 시간대에 맞는 값 넣어주기
    for hour in range(24):
      if hour == 24:
        hour = 0
      time = hour * 60
      value = list(tem_df[tem_df['time'] == time][tem])[0]
      temp_date[f'{hour}'] = value
    
    # 만들어진 df 합치기
    temp_point = temp_point.append(temp_date)
    date += timedelta(days=1)

  # 만들어진 temp_point 를 가지고, PCA를 수행한다.
  x = temp_point.drop(['date', "min", "max"], axis=1).values # 독립변인들의 value값만 추출
  x = StandardScaler().fit_transform(x) # x객체에 x를 표준화한 데이터를 저장
  
  pca = PCA(n_components=24) # 주성분을 몇개로 할지 결정
  printcipalComponents = pca.fit_transform(x)

  # eigen value를 통해, 몇개의 축을 선정할 것인지 판단하자.
  print('explained variance ratio :', pca.explained_variance_ratio_)

  # 주성분으로 이루어진 데이터 프레임 구성
  pca_df = pd.DataFrame(data= printcipalComponents[:,:2], columns=["pc1", "pc2"])

  # index 초기화
  temp_point = temp_point.reset_index()
  pca_df = pca_df.reset_index()

  # df 합치기
  temp_point = pd.merge(temp_point, pca_df, left_index=True, right_index=True, how='left')
  temp_point = temp_point.drop(["index_x", "index_y"], axis=1)
  # print(temp_point)

  # clustering을 위한 df
  cluster_df = temp_point[["pc1", "pc2", "min", "max"]]
  # print(cluster_df)

  kmeans = KMeans(n_clusters=k)
  kmeans.fit(cluster_df)
  temp_point[target+'-point'] = kmeans.predict(cluster_df)
  temp_point['plant'] = plant
  temp_point['loc'] = loc
  #print(temp_point[['date', 'plant', 'loc', target + '-point']])
  return temp_point[['date', 'plant', 'loc', target + '-point']]

# point 만들기
for tem in tem_name:
  cluster_df[tem + '-point'] = -1
for tem in tem_name:
  data_tem = data[['date', 'plant', 'loc', 'time', tem]]
  for plant in range(1, 3):
    for loc in range(1, 4):
      data_slice = data_tem[(data_tem['plant'] == plant) & (data_tem['loc'] == loc)]
      data_slice = point_fn(data_slice, plant, loc, tem)
      temp_cluster_df = cluster_df[(cluster_df['plant'] == plant) & (cluster_df['loc'] == loc)]
      temp_cluster_df[tem + '-point'] =  list(data_slice[tem + '-point'])
      cluster_df.loc[temp_cluster_df.index] = temp_cluster_df
cluster_df

cluster_df

## in, OUT cluster 파일 저장하기

cluster_df.to_csv(os.path.join(DATA_CONV_PATH, "date_trend_point.csv"))
