==> 4_code/main/18_PRED_TEST.ipynb <==
# 해당 파일 설명

이 코드를 실행하기 위한 파일, /convert 경로에 다음의 파일이 필요합니다.

1. plant_test_all_weather_feature_add_cluster_add_out_in.csv
2. date_trend_point.csv

이 코드를 실행하기 위한 파일, /model 경로에 다음의 파일이 필요합니다.

1. model_cond_all_es1000.pkl

이 코드를 모두 실행하고 나면, /result 경로에 다음의 파일이 생성됩니다.

1. es1000_th_08.csv

## 추가 설명

test 데이터를 model_cond_all_es1000 을 통해 cond_loc를 예측하여

그 결과를 제출형식과 맞게 만들어 result폴더로 저장합니다.

# 1. Google drive Mount

from google.colab import auth
auth.authenticate_user()

from google.colab import drive
drive.mount('/content/gdrive')

# 2. Define Directory

# 사용하기 전 변경해주어야 하는 파라미터
# BASE_WORKING_DIR = "/content/gdrive/My Drive/weather_bigcontest/4_code/"
BASE_WORKING_DIR = "/content/gdrive/My Drive/weather_bigcontest/4_code/"
USERNAME = "park"

import os
# 사용하는 디렉토리 정의
DATA_BASE_PATH = os.path.join(BASE_WORKING_DIR, "data")
DATA_RAW_PATH = os.path.join(DATA_BASE_PATH, "raw")
DATA_EXPORT_PATH = os.path.join(DATA_BASE_PATH, "export")
DATA_CONV_USER_PATH = os.path.join(DATA_BASE_PATH, USERNAME)
DATA_CONV_PATH = os.path.join(DATA_BASE_PATH, "converted")
MODEL_PATH = os.path.join(BASE_WORKING_DIR, "model")
RESULT_PATH = os.path.join(BASE_WORKING_DIR, "result")

WORKING_DIR = os.path.join(BASE_WORKING_DIR, USERNAME)

print('Current Work Directory :',  os.getcwd(), '\n')
print('Username :',  USERNAME, '\n')

print("DATA_BASE_PATH : ", DATA_BASE_PATH)
print("DATA_RAW_PATH : ", DATA_RAW_PATH)
print("DATA_EXPORT_PATH : ", DATA_EXPORT_PATH)
print("DATA_CONV_USER_PATH : ", DATA_CONV_USER_PATH)
print("DATA_CONV_PATH : ", DATA_CONV_PATH)

print("MODEL_PATH : ", MODEL_PATH)

print("RESULT_PATH : ", RESULT_PATH)

print("WORKING_DIR : ", WORKING_DIR)


# 데이터 무엇이 있는지 확인하자.
# 아래 코드를 수행하게 되면, "../input/"에서 리눅스 명령어 "ls"를 수행한 결과를 보여주게 된다. 이 때 나온 binary code를 "utf8"로 디코딩해서 보여준다.
from subprocess import check_output
print("===== Raw Data List =====\n", check_output(["ls", DATA_RAW_PATH]).decode("utf8"))
print("===== Export Data List =====\n" ,check_output(["ls", DATA_EXPORT_PATH]).decode("utf8"))
print("===== Converted User Data List =====\n" ,check_output(["ls", DATA_CONV_USER_PATH]).decode("utf8"))
print("===== Converted Data List =====\n" ,check_output(["ls", DATA_CONV_PATH]).decode("utf8"))
print("===== Model List =====\n" ,check_output(["ls", MODEL_PATH]).decode("utf8"))
print("===== Result file List =====\n" ,check_output(["ls", RESULT_PATH]).decode("utf8"))

# 4. Import Library

# 기본 데이터 라이브러리 로드
import sys # 시스템 파라미터에 접근할 수 있게 도와준다.
print("Python version: {}". format(sys.version))

import pandas as pd # 데이터 정제에 도움을 주는 라이브러리
print("pandas version: {}". format(pd.__version__))

import matplotlib # 매트랩에서 사용하는 시각화 도구를 사용할 수 있게 도와주는 시각화 도구
print("matplotlib version: {}". format(matplotlib.__version__))

import numpy as np # 행렬 계산을 위해 필요한 라이브러리
print("NumPy version: {}". format(np.__version__))

import scipy as sp # 수학 관련 함수가 내장된 라이브러리
print("SciPy version: {}". format(sp.__version__)) 

import IPython
from IPython import display # 주피터 노트북에서 예쁘게 시각화 해주는 도구
print("IPython version: {}". format(IPython.__version__)) 

import sklearn # 각종 통계 도구와 머신 러닝 알고리즘이 내장되어 있는 라이브러리
print("scikit-learn version: {}". format(sklearn.__version__))

# 파이썬 내장 라이브러리
import random
import time
import os
from datetime import datetime, timedelta
from dateutil.relativedelta import relativedelta
import copy


# Jupyter Notebook 이나 ipython 을 사용하다보면 향후 버전이 올라갈 때 변경될 사항 등을 알려주는 경고 메시지(warning message)를 뜨지 않게 해준다.
import warnings
warnings.filterwarnings('ignore')
print('-'*25)

# 모델링 라이브러리, 수학 계산 툴 로드
# 일반적인 모델링 라이브러리
from sklearn import svm, tree, linear_model, neighbors, naive_bayes, ensemble, discriminant_analysis, gaussian_process
from xgboost import XGBClassifier
import lightgbm

# 모델링 시 헬퍼 함수들
from sklearn.preprocessing import OneHotEncoder, LabelEncoder
from sklearn import feature_selection
from sklearn import model_selection
from sklearn import metrics

# interpolation 함수
import scipy.interpolate as spi

# model load
import pickle
from sklearn.externals import joblib

# 시각화 도구
import matplotlib as mpl
import matplotlib.pyplot as plt
import matplotlib.pylab as pylab
import seaborn as sns
from pandas.plotting import scatter_matrix

# 시각화 도구 default 세팅
# 주피터 노트북에서 plot 결과를 볼 수 있게 해준다.
%matplotlib inline
mpl.style.use('ggplot') # matplotlib에서 plot되는 결과를 선택할 수 있다.
sns.set_style('white') # seaborn에서 사용할 style을 설정할 수 있다.
pylab.rcParams['figure.figsize'] = 12,8 # plot의 크기와 선 등의 기본 값을 설정할 수 있다.

# 시각화 Style 설정하기
# print(plt.style.available)
# ['bmh', 'classic', 'dark_background', 'fast', 'fivethirtyeight', 'ggplot', 'grayscale', 'seaborn-bright', 'seaborn-colorblind', 'seaborn-dark-palette', 'seaborn-dark', 'seaborn-darkgrid', 'seaborn-deep', 'seaborn-muted', 'seaborn-notebook', 'seaborn-paper', 'seaborn-pastel', 'seaborn-poster', 'seaborn-talk', 'seaborn-ticks', 'seaborn-white', 'seaborn-whitegrid', 'seaborn', 'Solarize_Light2', 'tableau-colorblind10', '_classic_test']

plt.style.use(['fivethirtyeight'])
sns.set_style('darkgrid')

# 5. Get the Data

## 1. plant data 가져오기

# plant 데이터 
#data_train = pd.read_csv(os.path.join(DATA_CONV_PATH, "plant_train_all_weather_feature_add_cluster.csv"))
data_test = pd.read_csv(os.path.join(DATA_CONV_PATH, "plant_test_all_weather_feature_add_cluster_add_out_in.csv"))

# unnamed: 0 제거
#data_train.drop(['Unnamed: 0'], axis='columns', inplace=True)
data_test.drop(['Unnamed: 0'], axis='columns', inplace=True)

# date 변수 str -> datetime.date()
#data_train['date'] = data_train['date'].apply(lambda x : datetime.strptime(x + " 00:00:00", '%Y-%m-%d %H:%M:%S').date()) 
data_test['date'] = data_test['date'].apply(lambda x : datetime.strptime(x + " 00:00:00", '%Y-%m-%d %H:%M:%S').date()) 

data_test

# date 별로 in, out 클러스터링 데이터
data_cluster = pd.read_csv(os.path.join(DATA_CONV_PATH, "date_trend_point.csv"))

data_cluster.drop(['Unnamed: 0'], axis='columns', inplace=True)

data_cluster['date'] = data_cluster['date'].apply(lambda x : datetime.strptime(x + " 00:00:00", '%Y-%m-%d %H:%M:%S').date()) 

data_cluster

data_test = pd.merge(data_test, data_cluster, how='left', on = ['date', 'plant', 'loc'])
data_test

# Modeling


# 예측에 필요한 변수 정의
data_columns = list(data_test.columns)
remove_list = ['mea_ddhr', 'date']
for rmv in remove_list:
  data_columns.remove(rmv)
data_columns

# 예측에 필요한 변수 정의 --> 순서 까지 포함
data_columns = ['plant',
 'loc',
 'tem_in',
 'hum_in',
 'tem_coil',
 'tem_out_loc',
 'hum_out_loc',
 'time',
 'dangjin-ta',
 'dangjin-hm',
 'dangjin-ws',
 'dangjin-rn',
 'sinpyeong-ta',
 'sinpyeong-ws',
 'sinpyeong-rn',
 'sinpyeong-lp',
 'sinpyeong-ap',
 'seosan-ta',
 'seosan-hm',
 'seosan-ws',
 'seosan-rn',
 'seosan-dew',
 'seosan-sp',
 'seosan-lp',
 'seosan-ap',
 'seosan-sr',
 'seosan-ss',
 'seosan-ground',
 'songsan-ta',
 'songsan-hm',
 'songsan-ws',
 'songsan-rn',
 'songak-ta',
 'songak-hm',
 'songak-ws',
 'songak-rn',
 'seosan-ta-trend',
 'seosan-hm-trend',
 'seosan-ground-trend',
 'seosan-ta-point',
 'seosan-hm-point',
 'seosan-ground-point',
 'seosan-ta-changed',
 'seosan-hm-changed',
 'seosan-ground-changed',
 'tem_out_loc-trend',
 'hum_out_loc-trend',
 'tem_in-trend',
 'hum_in-trend',
 'tem_coil-trend',
 'tem_out_loc-point',
 'hum_out_loc-point',
 'tem_in-point',
 'hum_in-point',
 'tem_coil-point']
data_columns

# cond_loc 예측 모델 가지고 오기
model_cond = joblib.load(os.path.join(MODEL_PATH, "model_cond_all_es1000.pkl"))
model_cond.keys()

# 예측하기
data_test_cond = data_test[['mea_ddhr', 'plant', 'loc']]

for key in model_cond.keys():
  model = model_cond[key]
  data_test_cond[key[0]] = model.predict_proba(data_test[data_columns])[:, 1]
data_test_cond

data_test_cond.columns = ['mea_ddhr', 'plant', 'loc', 'cond_loc']

def makeTestSubmitDateFormatEqual(df):
  """
  describe : test 의 mea_ddhr을 submit의 mea_ddhr의 양식과 같게 만드는 함수

  <input>
  df = pd.dataframe -> test 데이터

  <output>
  df = pd.dataframe -> test 데이터의 mea_ddhr이 변환된 df
  """
  ret_df = copy.deepcopy(df)
  for i in range(ret_df.shape[0]):
    ddhr = ret_df.loc[i, "mea_ddhr"]
    date = ddhr.split()[0]
    times = (ddhr.split()[1]).split(":")

    hour = str(int(times[0]))
    minute = times[1]
    new_time = hour + ":" + minute
    ret_df.loc[i, "mea_ddhr"] = date + " " + new_time
  return ret_df

data_predicted = makeTestSubmitDateFormatEqual(data_test_cond)
data_predicted

# submit load
data_submit = pd.read_csv(os.path.join(DATA_RAW_PATH, "submit.csv"))
data_submit.drop(['X24H_COND_LOC','X24H_COND_LOC_PROB','X48H_COND_LOC','X48H_COND_LOC_PROB'], axis='columns', inplace=True)
# data_submit = stringToDatetime(data_submit, 'X24H_TMA')
# data_submit = stringToDatetime(data_submit, 'X48H_TMA')
data_submit['X24H_TMA'] = data_submit['X24H_TMA'].apply(lambda x : str(x))
data_submit['X48H_TMA'] = data_submit['X48H_TMA'].apply(lambda x : str(x))
data_submit

data_predicted.columns = ['X24H_TMA','PLANT','LOC','X24H_COND_LOC_PROB']
data_predicted

data_submit = pd.merge(data_submit, data_predicted, how = 'left', on=['X24H_TMA', 'PLANT', 'LOC'])
data_submit

data_predicted.columns = ['X48H_TMA','PLANT','LOC','X48H_COND_LOC_PROB']
data_predicted

data_submit = pd.merge(data_submit, data_predicted, how = 'left', on=['X48H_TMA', 'PLANT', 'LOC'])
data_submit

threshold = 0.08

data_submit['X24H_COND_LOC'] = (data_submit['X24H_COND_LOC_PROB'] >= threshold).apply(lambda x : 1 if x == True else 0)
print("24시간 이후 예측한 true의 개수는 :", sum(data_submit['X24H_COND_LOC'] == 1) , "개 입니다.")

data_submit['X48H_COND_LOC'] = (data_submit['X48H_COND_LOC_PROB'] >= threshold).apply(lambda x : 1 if x == True else 0)
print("48시간 이후 예측한 true의 개수는 :", sum(data_submit['X48H_COND_LOC'] == 1) , "개 입니다.")

for threshold in np.arange(0.06, 0.1, 0.001):
  threshold = round(threshold,4)
  data_submit['X24H_COND_LOC'] = (data_submit['X24H_COND_LOC_PROB'] >= threshold).apply(lambda x : 1 if x == True else 0)
  data_submit['X48H_COND_LOC'] = (data_submit['X48H_COND_LOC_PROB'] >= threshold).apply(lambda x : 1 if x == True else 0)
  print(threshold, sum(data_submit['X24H_COND_LOC'] == 1), sum(data_submit['X48H_COND_LOC'] == 1))

# 제출 양식에 맞게 column을 재정렬한다.
data_submit = data_submit[["MEA_DDHR", "PLANT",	"LOC", "X24H_TMA", "X24H_COND_LOC",	"X24H_COND_LOC_PROB",	"X48H_TMA", "X48H_COND_LOC",	"X48H_COND_LOC_PROB"]]
data_submit

data_submit['X24H_COND_LOC_PROB'] = data_submit['X24H_COND_LOC_PROB']*100
data_submit['X48H_COND_LOC_PROB'] = data_submit['X48H_COND_LOC_PROB']*100

data_submit

data_submit.to_csv(os.path.join(RESULT_PATH, 'es1000_th_08.csv'))
