==> 4_code/main/24_Forecast_to_weather.ipynb <==
# 해당 파일 설명
이 코드를 실행하기 위한 파일, /convert 경로에 다음의 파일이 필요합니다.

1. ALL_FC_DATA_MEAN.csv
2. data_all_time_out_in.csv

이 코드를 모두 실행하고 나면, /convert 경로에 다음의 파일이 생성됩니다.

1. ALL_FC_DATA_MEAN_ADD_WEATHER.csv

이 코드를 모두 실행하고 나면, /model 경로에 다음의 파일이 생성됩니다.

1. model_fc_all.pkl

## 추가 설명
예보 데이터를 기반으로 weather 데이터를 예측합니다. 

이때 굉장히 높은 정확도를 갖는 모델을 만들었습니다.

그 모델을 model_fc_all.pkl 로 저장합니다.

# 1. Google Drive Mount

from google.colab import auth
auth.authenticate_user()

from google.colab import drive
drive.mount('/content/gdrive')

# 2. Define Directory

# 사용하기 전 변경해주어야 하는 파라미터
BASE_WORKING_DIR = "/content/gdrive/My Drive/weather_bigcontest/4_code/"
USERNAME = "park"

import os
# 사용하는 디렉토리 정의
DATA_BASE_PATH = os.path.join(BASE_WORKING_DIR, "data")
DATA_RAW_PATH = os.path.join(DATA_BASE_PATH, "raw")
DATA_EXPORT_PATH = os.path.join(DATA_BASE_PATH, "export")
DATA_CONV_USER_PATH = os.path.join(DATA_BASE_PATH, USERNAME)
DATA_CONV_PATH = os.path.join(DATA_BASE_PATH, "converted")
MODEL_PATH = os.path.join(BASE_WORKING_DIR, "model")
RESULT_PATH = os.path.join(BASE_WORKING_DIR, "result")

WORKING_DIR = os.path.join(BASE_WORKING_DIR, USERNAME)

print('Current Work Directory :',  os.getcwd(), '\n')
print('Username :',  USERNAME, '\n')

print("DATA_BASE_PATH : ", DATA_BASE_PATH)
print("DATA_RAW_PATH : ", DATA_RAW_PATH)
print("DATA_EXPORT_PATH : ", DATA_EXPORT_PATH)
print("DATA_CONV_USER_PATH : ", DATA_CONV_USER_PATH)
print("DATA_CONV_PATH : ", DATA_CONV_PATH)

print("MODEL_PATH : ", MODEL_PATH)

print("RESULT_PATH : ", RESULT_PATH)

print("WORKING_DIR : ", WORKING_DIR)

# 데이터 무엇이 있는지 확인하자.
# 아래 코드를 수행하게 되면, "../input/"에서 리눅스 명령어 "ls"를 수행한 결과를 보여주게 된다. 이 때 나온 binary code를 "utf8"로 디코딩해서 보여준다.
from subprocess import check_output
print("===== Raw Data List =====\n", check_output(["ls", DATA_RAW_PATH]).decode("utf8"))
print("===== Export Data List =====\n" ,check_output(["ls", DATA_EXPORT_PATH]).decode("utf8"))
print("===== Converted Data List =====\n" ,check_output(["ls", DATA_CONV_PATH]).decode("utf8"))

# 4. Import Library

# 기본 데이터 라이브러리 로드
import sys # 시스템 파라미터에 접근할 수 있게 도와준다.
print("Python version: {}". format(sys.version))

import pandas as pd # 데이터 정제에 도움을 주는 라이브러리
print("pandas version: {}". format(pd.__version__))

import matplotlib # 매트랩에서 사용하는 시각화 도구를 사용할 수 있게 도와주는 시각화 도구
print("matplotlib version: {}". format(matplotlib.__version__))

import numpy as np # 행렬 계산을 위해 필요한 라이브러리
print("NumPy version: {}". format(np.__version__))

import scipy as sp # 수학 관련 함수가 내장된 라이브러리
print("SciPy version: {}". format(sp.__version__)) 

import IPython
from IPython import display # 주피터 노트북에서 예쁘게 시각화 해주는 도구
print("IPython version: {}". format(IPython.__version__)) 

import sklearn # 각종 통계 도구와 머신 러닝 알고리즘이 내장되어 있는 라이브러리
print("scikit-learn version: {}". format(sklearn.__version__))

# 파이썬 내장 라이브러리
import random
import time
import os


# Jupyter Notebook 이나 ipython 을 사용하다보면 향후 버전이 올라갈 때 변경될 사항 등을 알려주는 경고 메시지(warning message)를 뜨지 않게 해준다.
import warnings
warnings.filterwarnings('ignore')
print('-'*25)

# 모델링 라이브러리, 수학 계산 툴 로드
# 일반적인 모델링 라이브러리
from sklearn import svm, tree, linear_model, neighbors, naive_bayes, ensemble, discriminant_analysis, gaussian_process
from xgboost import XGBClassifier
from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler  # 표준화 패키지 라이브러리 

# 모델링 시 헬퍼 함수들
from sklearn.preprocessing import OneHotEncoder, LabelEncoder
from sklearn import feature_selection
from sklearn import model_selection
from sklearn import metrics
from sklearn.datasets import make_blobs
from sklearn.cluster import KMeans

# interpolation 함수
import scipy.interpolate as spi

# model load
import pickle
from sklearn.externals import joblib

# 시각화 도구
import matplotlib as mpl
import matplotlib.pyplot as plt
import matplotlib.pylab as pylab
import seaborn as sns
from pandas.plotting import scatter_matrix

# 시각화 도구 default 세팅
# 주피터 노트북에서 plot 결과를 볼 수 있게 해준다.
%matplotlib inline
mpl.style.use('ggplot') # matplotlib에서 plot되는 결과를 선택할 수 있다.
sns.set_style('white') # seaborn에서 사용할 style을 설정할 수 있다.
pylab.rcParams['figure.figsize'] = 12,8 # plot의 크기와 선 등의 기본 값을 설정할 수 있다.

# 시각화 Style 설정하기
# print(plt.style.available)
# ['bmh', 'classic', 'dark_background', 'fast', 'fivethirtyeight', 'ggplot', 'grayscale', 'seaborn-bright', 'seaborn-colorblind', 'seaborn-dark-palette', 'seaborn-dark', 'seaborn-darkgrid', 'seaborn-deep', 'seaborn-muted', 'seaborn-notebook', 'seaborn-paper', 'seaborn-pastel', 'seaborn-poster', 'seaborn-talk', 'seaborn-ticks', 'seaborn-white', 'seaborn-whitegrid', 'seaborn', 'Solarize_Light2', 'tableau-colorblind10', '_classic_test']

plt.style.use(['fivethirtyeight'])
sns.set_style('darkgrid')

# 5. Read Raw Data

FC_DATA = pd.read_csv(os.path.join(DATA_CONV_PATH, 'ALL_FC_DATA_MEAN.csv'))
FC_DATA.drop(columns=['Unnamed: 0'], inplace=True)
FC_DATA['time'] = FC_DATA['time'].apply(lambda x : int(x))
FC_DATA

data = pd.read_csv(os.path.join(DATA_CONV_PATH, 'data_all_time_out_in.csv'))
data.drop(columns=['Unnamed: 0'], inplace=True)
data['time'] = data['time'].apply(lambda x : int(x))
data

columns = ['date', 'time', 'dangjin-ta', 'dangjin-hm',
       'dangjin-ws', 'dangjin-rn', 'sinpyeong-ta', 'sinpyeong-ws',
       'sinpyeong-rn', 'sinpyeong-lp', 'sinpyeong-ap', 'seosan-ta',
       'seosan-hm', 'seosan-ws', 'seosan-rn', 'seosan-dew', 'seosan-sp',
       'seosan-lp', 'seosan-ap', 'seosan-sr', 'seosan-ss', 'seosan-ground',
       'songsan-ta', 'songsan-hm', 'songsan-ws', 'songsan-rn', 'songak-ta',
       'songak-hm', 'songak-ws', 'songak-rn']

data = data[columns]

FC_DATA = pd.merge(FC_DATA, data, how='left', on=['date', 'time'])

FC_DATA = FC_DATA.dropna()

FC_DATA.to_csv(os.path.join(DATA_CONV_PATH, 'ALL_FC_DATA_MEAN_ADD_WEATHER.csv'))

MLA = [
       #ensemble.AdaBoostRegressor(),
       #ensemble.BaggingRegressor(),
       ensemble.ExtraTreesRegressor(),
       #ensemble.GradientBoostingRegressor(),
       #ensemble.RandomForestRegressor(),

       #gaussian_process.GaussianProcessClassifier(),

       #linear_model.LogisticRegressionCV(),
       #linear_model.PassiveAggressiveClassifier(),
       #linear_model.RidgeClassifierCV(),
       #linear_model.SGDClassifier(),
       #linear_model.Perceptron(),

       #naive_bayes.BernoulliNB(),
       #naive_bayes.GaussianNB(),

       #svm.SVC(probability=True),
       #svm.NuSVC(probability=True),
       #svm.LinearSVC(),

       #tree.DecisionTreeClassifier(),
       #tree.ExtraTreeClassifier(),

       #discriminant_analysis.LinearDiscriminantAnalysis(),
       #discriminant_analysis.QuadraticDiscriminantAnalysis(),

       #XGBClassifier()
       #lightgbm.LGBMClassifier()
]
MLA

data_columns = ['time', 'songsan_ta', 'songsan_hm', 'songsan_ws', 'songak_ta',
       'songak_hm', 'songak_ws', 'sinpyeong_ta', 'sinpyeong_hm',
       'sinpyeong_ws', 'dangjin1_ta', 'dangjin1_hm', 'dangjin1_ws',
       'dangjin2_ta', 'dangjin2_hm', 'dangjin2_ws', 'dangjin3_ta',
       'dangjin3_hm', 'dangjin3_ws', 'godae_ta', 'godae_hm', 'godae_ws',
       'emam_ta', 'emam_hm', 'emam_ws', 'susuk_ta', 'susuk_hm', 'susuk_ws',
       'suknam_ta', 'suknam_hm', 'suknam_ws']

Targets = ['seosan-ta',
       'seosan-hm', 'seosan-ws', 'seosan-rn', 'seosan-dew', 'seosan-sp',
       'seosan-lp', 'seosan-sr', 'seosan-ss', 'seosan-ground',
       'songsan-ta', 'songsan-hm', 'songsan-ws']

model = dict()

import copy

Targets = Targets

data_train_bin = data_columns

cv_split = model_selection.ShuffleSplit(n_splits=10, test_size=.3, train_size=.6, random_state=0)

MLA_columns = ['MLA Name', 'MLA Parameters', 'MLA Test Accuracy Mean', 'MLA Test Accuracy 3*STD', 'MLA Time']

MLA_compares = list()
MLA_predict = dict()
mla_index = 0
for Target in Targets:
  MLA_compare = pd.DataFrame(columns=MLA_columns)
  MLA_compares.append(MLA_compare)


  row_index = 0
  for alg in MLA:
    print(alg)
    data = copy.deepcopy(FC_DATA)
    MLA_name = alg.__class__.__name__
    MLA_compares[mla_index].loc[row_index, 'MLA Name'] = MLA_name
    MLA_compares[mla_index].loc[row_index, 'MLA Parameters'] = str(alg.get_params())

    cv_results = model_selection.cross_validate(alg, data[data_train_bin], data[Target], cv=cv_split)
    MLA_compares[mla_index].loc[row_index, 'MLA Time'] = cv_results['fit_time'].mean()
    MLA_compares[mla_index].loc[row_index, 'MLA Test Accuracy Mean'] = cv_results['test_score'].mean()
    MLA_compares[mla_index].loc[row_index, 'MLA Test Accuracy 3*STD'] = cv_results['test_score'].std()*3

    alg.fit(data[data_train_bin], data[Target])
    MLA_predict[(MLA_name, Target)] = alg.predict(data[data_train_bin])
    model[Target] = copy.deepcopy(alg)
    row_index += 1
  MLA_compares[mla_index].sort_values(by = ['MLA Test Accuracy Mean'], ascending = False, inplace = True)
  mla_index += 1 

for idx in range(len(Targets)):
  print(MLA_compares[idx])

print('seosan-ta')
MLA_compares[0]

print('seosan-hm')
MLA_compares[1]

print('seosan-ws')
MLA_compares[2]

print('seosan-rn')
MLA_compares[3]

print('seosan-dew')
MLA_compares[4]

print('seosan-sp')
MLA_compares[5]

print('seosan-lp')
MLA_compares[6]

print('seosan-sr')
MLA_compares[7]

print('seosan-ss')
MLA_compares[8]

print('seosan-ground')
MLA_compares[9]

print('songsan-ta')
MLA_compares[10]

print('songsan-hm')
MLA_compares[11]

print('songsan-ws')
MLA_compares[12]

joblib.dump(model, os.path.join(MODEL_PATH, "model_fc_all.pkl"))
